Moved public headers out of internals
