+++ AppCenter/AppCenter/MSChannelGroupProtocol.h
