+++ AppCenter/AppCenter/MSChannelProtocol.h
