+++ AppCenter/AppCenter/MSLog.h
