+++ AppCenterAnalytics/AppCenterAnalytics/Model/MSEventLog.h
