+++ AppCenterAnalytics/AppCenterAnalytics/Model/MSEventLog.m
