+++ AppCenterAnalytics/AppCenterAnalytics/Model/MSLogWithNameAndProperties.h
