+++ AppCenterAnalytics/AppCenterAnalytics/Model/MSLogWithNameAndProperties.m
