+++ AppCenterCrashes/AppCenterCrashes/Model/MSErrorReport.h
