+++ AppCenterDistribute/AppCenterDistribute/Model/MSReleaseDetails.h
