+++ AppCenterDistribute/AppCenterDistribute/Model/MSReleaseDetails.m
